Move common bits to make them reusable from the Scout CLI
